+++ CHANGELOG.md
@@ -21,6 +21,7 @@
 - Change the minimum requirements to be lower.
 - Change Wine GE to be installed as part of the first-time setup.
 - Add support for VMware Fusion and VMware Workstation.
+- Change ZeroTier VPN to not be enabled by default.
 
 ## 3.1.1 - The Boot Fix Update